feat: change name from cv-btn to cv-download and rebuild the design for the button. Import the new component RotatingTextCircle
